refactor: move migrations to sep file
