+++ runtime/centrifuge/src/lib.rs
@@ -93,6 +93,7 @@ use xcm_executor::XcmExecutor;
 
 use crate::xcm::{XcmConfig, XcmOriginToTransactDispatchOrigin};
 
+mod migrations;
 mod weights;
 pub mod xcm;
 
@@ -1167,7 +1168,7 @@ frame_support::parameter_types! {
 	pub const SingleCurrencyMovement: u32 = 1;
 	#[derive(scale_info::TypeInfo, Debug, PartialEq, Eq, Clone)]
 	pub const MaxChangesPerEpoch: u32 = 50;
-	pub const RewardsPalletId: PalletId = PalletId(*b"d/reward");
+	pub const RewardsPalletId: PalletId = PalletId(cfg_primitives::constants::PALLET_ID_BLOCK_REWARDS);
 	pub const RewardCurrency: CurrencyId = CurrencyId::Native;
 }
 
@@ -1739,26 +1740,6 @@ pub type UncheckedExtrinsic =
 /// Extrinsic type that has already been checked.
 pub type CheckedExtrinsic = generic::CheckedExtrinsic<AccountId, RuntimeCall, SignedExtra>;
 
-parameter_types! {
-	// = 16.65 CFG per epoch (12h)
-	// https://gov.centrifuge.io/t/rfc-add-a-block-reward-and-improve-collators-cycle/4766
-	pub const CollatorRewards: Balance = 8_325 * MILLI_CFG;
-	// = 20,096 CFG per epoch (12h)
-	pub const TotalRewards: Balance = 10_048 * CFG;
-}
-
-type UpgradeCentrifuge1019 = (
-	pallet_loans_ref::migrations::v1::Migration<Runtime>,
-	crate::session_key_migration::ExtendSessionKeys,
-	pallet_block_rewards::migrations::InitBlockRewards<Runtime, CollatorRewards, TotalRewards>,
-	pallet_rewards::migrations::new_instance::FundExistentialDeposit<
-		Runtime,
-		pallet_rewards::Instance1,
-		NativeToken,
-		ExistentialDeposit,
-	>,
-);
-
 /// Executive: handles dispatch to the various modules.
 pub type Executive = frame_executive::Executive<
 	Runtime,
@@ -1766,104 +1747,8 @@ pub type Executive = frame_executive::Executive<
 	frame_system::ChainContext<Runtime>,
 	Runtime,
 	AllPalletsWithSystem,
-	UpgradeCentrifuge1019,
+	crate::migrations::UpgradeCentrifuge1019,
 >;
-
-// Session Key migration required for BlockRewards
-mod session_key_migration {
-	use frame_support::{pallet_prelude::*, storage_alias};
-
-	use super::*;
-	pub struct ExtendSessionKeys;
-
-	impl_opaque_keys! {
-		pub struct OldSessionKeys {
-			pub aura: Aura,
-		}
-	}
-
-	#[storage_alias]
-	type QueuedKeys<T: pallet_session::Config> = StorageValue<
-		pallet_session::Pallet<T>,
-		// hex_literal::hex!("cec5070d609dd3497f72bde07fc96ba0"),
-		Vec<(AccountId, OldSessionKeys)>,
-		ValueQuery,
-	>;
-
-	impl frame_support::traits::OnRuntimeUpgrade for ExtendSessionKeys {
-		#[cfg(feature = "try-runtime")]
-		fn pre_upgrade() -> Result<Vec<u8>, &'static str> {
-			let queued: Vec<(AccountId, OldSessionKeys)> = QueuedKeys::<Runtime>::get();
-
-			assert!(
-				!queued.len().is_zero(),
-				"Queued session keys must never be empty"
-			);
-
-			Ok(queued.encode())
-		}
-
-		fn on_runtime_upgrade() -> frame_support::weights::Weight {
-			use sp_runtime::SaturatedConversion;
-
-			// We must not bump the storage version of pallet_session to stay in sync with
-			// Substrate. As this migration is tightly coupled to the initialization of
-			// BlockRewards, it should be sane to trigger the session migration based on the
-			// BlockRewards storage version.
-			if BlockRewards::on_chain_storage_version() == StorageVersion::new(0) {
-				Session::upgrade_keys::<OldSessionKeys, _>(|_, old_key| SessionKeys {
-					aura: old_key.aura.clone(),
-					block_rewards: old_key.aura,
-				});
-
-				let n: u64 = pallet_session::Validators::<Runtime>::get()
-					.len()
-					.saturated_into();
-
-				<Runtime as frame_system::Config>::DbWeight::get()
-					.reads_writes(n.saturating_add(1), n)
-			} else {
-				<Runtime as frame_system::Config>::DbWeight::get().reads(1)
-			}
-		}
-
-		#[cfg(feature = "try-runtime")]
-		fn post_upgrade(pre_state: Vec<u8>) -> Result<(), &'static str> {
-			let pre_state: Vec<(AccountId, OldSessionKeys)> =
-				Decode::decode(&mut pre_state.as_slice())
-					.expect("Pre state parameter should be generated by pre-upgrade hooks");
-
-			let queued: Vec<(AccountId, SessionKeys)> =
-				pallet_session::QueuedKeys::<Runtime>::get();
-
-			assert!(
-				pre_state.len() == queued.len(),
-				"Mismatch in number of session keys: pre {} vs post {}",
-				pre_state.len(),
-				queued.len()
-			);
-
-			for i in 0..queued.len() {
-				assert!(
-					&queued[i].0 == &pre_state[i].0,
-					"Failed to match old validator key at index {}: old {:?} vs {:?} new",
-					i,
-					&pre_state[i].0,
-					&queued[i].0
-				);
-				assert!(
-					&queued[i].1.aura == &pre_state[i].1.aura,
-					"Failed to match old aura key at index {}. old {:?} vs {:?} new",
-					i,
-					&pre_state[i].1.aura,
-					&queued[i].1.aura,
-				);
-			}
-
-			Ok(())
-		}
-	}
-}
 
 #[cfg(not(feature = "disable-runtime-api"))]
 impl_runtime_apis! {
